papi/buildsys: add util macros to configure
Currently, when we need to test for some functionality locally we store
the current FLAG-VARIABLE's value in a temporary variable, change the
original's value to perform the local test, and eventually restore the
original value of the variable. This process is error prone as we might
forget to restore the variable or set it wrongly. Using AX_VAR_PUSHVALUE
and AX_VAR_POPVALUE macros, which store the original value, set the
variable to the desired value and restore it, respectively, is cleaner
and less subject to errors, as PUSH and POP macros go in pairs.

Another macro that is introduced by this commit is the AX_APPEND_FLAG
which avoids appending the same flag to a FLAG-VARIABLE if this already
contains it.