BuildSys: fix bogus check os OSVER in --with-OSVER
The intended behavior of the check is to set OSVER to `uname -r` if the
OS is either 'bgp' or 'bgq'. However, the check uses a or for the test.
The or shortcuts if the first logic statement is true setting OSVER to
`uname -r` even when OS=bgq. The fix is to use an 'and' rather than an
'or' in the test.