BuildSys: move libraries to LIBS and linker flags to LDFLAGS
Some of the code (e.g. the utility programs) need libpapi to be compiled
first. In order to do this the rules for making utils contain a
dependency on libpapi. The library can be compiled as static or shared.
To distinguish the two there is a flag variable called papiLIBS, which
is set to either 'static' or 'shared' at configuration time. This is
then exported and substituted into another variable (LIBS) in
Makefile.in. This is a gigantic hack. First LIBS is used for both
libraries to be passed to the linker and this switch flag. Second
autotools already knows how to manage dependencies without manually
specifying them.

The current mechanism is carried over in the code but LIBS has been
renamed to PAPILIBTYPE to convey this is a switch flag for what type of
library should be compiled.