RocmCmp: improve error message reporting text
Currently the rocm component returns a generic error when dlopen() fails
loading the requested cuda libraries. This is not very helpful when the
error is not due to the missing PAPI_ROCM_ROOT definition but rather to
some more specific problem in dlopen() not being able to do its job. For
example, if the dependency libraries needed by the dlopen()'ed library
are not found this error still falls in the more generic one returned by
PAPI. Adding dlerror() string to the PAPI message gives more useful
information that the user can use to fix the issue.