rocm: use HSA_TOOLS_LIB to load rocprofiler
HSA runtime uses HSA_TOOLS_LIB to load and initialize rocprofiler. It is
unwise to load rocprofiler in PAPI regardless of what HSA does. We may
end up in situations in which PAPI is using a not properly initialized
rocprofiler. Force PAPI ROCm component to also look for
librocprofiler64.so through the HSA_TOOLS_LIB variable. Also remove the
logic in which the component was setting up the HSA_TOOLS_LIB.