Add git ignore file
Currently we require users to build PAPI from within the source
directory. This causes all the files generated by the build process to
pollute the source directory of the repository. This also causes issue
in the development phase, during which frequent builds and changes are
pretty common.

When looking for changed files in the repo using git diff we are
presented a long list of files that are the result of the build process
and have nothing to do with the source code. This patch introduces a
.gitignore file that causes git to ignore all added files, whether
generated automatically or intended to be added by the user. New files
in this way have to be added using `git add -f -- filename`.