configure: get rid of os headers variables
Currently, we use a variable placeholder for context and lock header
files in the papi_internal.h and papi_lock.h files, respectively. The
variables in object are called OSCONTEXT and OSLOCK. These are resolved
at configure time and replaced with the name of the right header,
depending on the OS being used. This is way of selecting headers is
unusual and may prove difficult to transition to a more modern build
system in the future. The way configure typically resolves this type of
cases is to define a CPPFLAGS flag that selects the proper header from a
hard coded list of "if defined else ..." headers in the target file.