testlib: add papi event test interface
The papi event test interface allows to pick a defined number of events
from a component randomly, add those events to an eventset, instrument
and run an user defined kernel function. The function can be ran a
defined number of iterations and the standard deviation for each of the
events is computed. The standard deviation is compared against a user
defined standard deviation threshold and an error is returned if such
threshold is crossed.