Update eventset state in PAPI_stop even when a failure happens
Currently, if there is a failure in PAPI_stop the eventset state is not
updated from "running" to "stopped". This causes calls to cleanup and
destroy to fail and the consequent leaking of memory.

This patch fixes the problem by funneling all failure in PAPI_stop
through a single exit point that also cleans up the eventset state flags
whenever is appropriate.

This patch addresses issue #69