sysdetect: fix non-sequential numbering of POWER9 numas
Linux numbers POWER9 numa nodes non-sequentially. For example, if the
system has two sockets POWER9 with 64 hw threads each, the first half of
the cores will be associated to numa node0, while the second half will
be associated to numa node8 (instead of 1). Because most of the logic in
sysdetect relies on enumeration of numa nodes being sequential, this
inconsistency causes a segmentation fault. This patch is an easy fix
that bails out if the expected numa number is greater than the number of
detected numas. For POWER9 this causes only half of the hardware threads
affinity to be printed by papi_hardware_avail.