sysdetect: improve mpi checks in configure
MPI support is needed in sysdetect in order for processes to exchange
GPU affinity and allow users to know what threads can access the same
GPUs. Tests in sysdetect also require MPI support. The way the configure
used to check for MPI support availability was by looking for the mpi.h
header in the system paths. This built PAPI with MPI whenever the mpi
header was found, although users might have not requested it.

Instead, this patch, support MPI through configure time options:
 (1) --with-mpi: takes the path to the MPI installation
 (2) --with-mpi-include and --with-mpi-lib which take, respectively, the
     header directory and the lib directory.