rocm: add new open functions to the rocp layer
This patch is the second in a series of several patches that aims at
removing the need for a native event table in the case the user only
uses named events. In this patch a new open functions is introduced:
rocp_ctx_open_v2. This is almost identical to rocp_ctx_open except it
does not need a reference to the event table. This new open interface is
also replaced in the rocm.c layer.