rocm: rework verify_events function
The verify_events function makes sure that, in intercept mode, all
eventsets have the same events. This is dictated by rocprofiler that,
currently, does not allow resetting intercept mode callbacks. The way
this check is carried out is by going through a list of intercept events
(kept internally by the rocp layer) and a list of user requested events.
If the two differ, then there is a conflict and the new eventset cannot
be monitored. Use the htable to log the name of the intercept events and
check the presence/absence for conflict in verify_events.