rocm: sort events in the rocm layer instead of the rocp layer
The rocp layer needs events sorted by device. The sorting used to
happen in the rocp layer and counters would be eventually assigned
to the correct position in the user counter array. The same mechanism
is already available in the frontend through the ntv_info ni_position
attribute. Thus, do the sorting the in the rocm layer and set the
ni_position to the remapped event/counter. This allows to simplify
the rocp layer logic.