rocm: allow for different # events per device in eventset
The rocm component used to make some assumption when working in sampling
mode. More specifically, in the case of a single eventset containing
events from different devices, the component assumed every device had the
same number of events. This is a reasonable assumption to make because
in the typical case the user has a SIMD workload split across available
devices. It make sense therefore to monitor the same events on all
devices, in order to make apple to apple comparisons. This assumption,
however, does not allow for the case in which the user wants to monitor
different events on different devices. This might be the case for a MIMD
workload. This patch removes the limitation by allowing different number
of events to be monitored on different devices using a single eventset.